Refactor memory management and update build dependencies
Code refactoring was carried out to optimize memory management. The 'buddy_allocator.zig' was adjusted for commenting out 'dump' function for cleaner logs and 'freeInner' function to include 'idx' in the log message for better understanding. A 'zigavl' module is added to the 'compileKernelAction' function in 'build.zig', and a new key 'zigavl' is added in 'build.zig.zon' for dependency management. Changes were also made to 'pmm.zig' by implementing an AVL tree, and required de-initialize functions are provided for clean up.