Refactor GDT initialization and add GDTR structure
The commit involves refactoring the initialization of the Global Descriptor Table (GDT) for the x86_64 architecture. A GDTR structure is also added that provides the size and offset of GDT. The assembly function lgdt is introduced to load this GDTR. This helps improve the control over segment selectors and manages memory segmentation more efficiently.