Implemented basic Interrupt Descriptor Table (IDT) handling
In this commit, I implemented the basic handling of the Interrupt Descriptor Table (IDT) and updated the README.md. The basic exception handlers and Programmable Interrupt Controller (PIC) IRQs have been added, and the handling of interrupts has been improved. References to these changes have also been updated in the current development steps of the README.