Refactor debug logging and exception handling functions
Renamed the function `selfLogDebug` to `logDebugInfo` in the `gdt.zig` file to make it more intuitive. Also, refactored exception handling functions in `int.zig` by changing the name of `ExceptionFn` to `HandleFn`, which is more indicative of its purpose. Improved readability of a comment related to updating IDT with exceptions.