Refactor PCI driver, add BAR size and optimise code format
Refactor PCI driver code to include the size of the BAR and enhance readability. The changes also involved formatting adjustments, such as removal of excess spaces and rearranging of functions arguments to be on the same line. Some functionality was added to determine the amount of address space. Additionally, minor logic errors within the code were rectified.