Rename PatType to PATType in x86_64 paging module
The naming of 'PatType' and related methods in the x86_64 paging module have been revised to 'PATType' for better consistency and readability. Functions, variables, and instances where the old naming was used have been duly updated to reflect this change. The unrelated whitespace in 'Nvme.zig' file was also removed.