Implement enhanced interrupt handling and add NVMe interrupt handler
This commit significantly improves interrupt handling by changing `interruptWithAcknowledgeFnBind` to accept a `VectorIndex` and allowing it to process multiple handlers for a given interrupt vector. ISR handlers are now stored in a list with a new structure `ISRHandler`. The code also initiates and deinitializes this list during the kernel startup process. Furthermore, an interrupt handler for NVMe is established. A method for generating a unique identifier for PCI devices has also been introduced.