Refactor interrupt handling and exception binding
This commit refactors the existing interrupt handling and exception binding in the x86_64 architecture. The primary changes include renaming of functions to more appropriate ones like from exceptionFnBind to bindExceptionHandler and from interruptFnBind to bindIRQHandler. Moreover, the handling of function list for exceptions and interrupts has been streamlined with the replacement of initHandlerList, deinitHandlerList, and addHandler with their ISR equivalents - initISRMap, deinitISRMap, and addISR respectively.